mbewubot/agri_bot_model/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0 